Remove circular deps on Trie. #RDS-1694
