+++ crates/redstone/src/helpers/mod.rs
@@ -1,2 +1,3 @@
 pub mod hex;
 pub mod iter_into;
+pub mod trie;
