+++ crates/redstone/src/types/mod.rs
@@ -7,7 +7,6 @@ use alloc::vec::Vec;
 
 pub use bytes::Bytes;
 pub use feed_id::FeedId;
-pub use feed_id::Trie;
 pub use signer_address::SignerAddress;
 pub use timestamp_millis::TimestampMillis;
 pub use value::Value;
